Rename zeroconf to mdns_manager
